source/code/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.4
requests==2.31.0
gunicorn==21.2.0
email_validator==2.1.0
boto3==1.34.0
botocore==1.34.0
